Fix stubs and move some tests to http_common
