+++ Makefile.am
@@ -126,7 +126,6 @@ unittest_sources = \
 	tests/unittests/log/stub_log.c \
 	tests/unittests/database/stub_database.c \
 	tests/unittests/config/stub_accounts.c \
-	tests/unittests/tools/stub_http_common.c \
 	tests/unittests/tools/stub_http_upload.c \
 	tests/unittests/tools/stub_http_download.c \
 	tests/unittests/tools/stub_aesgcm_download.c \
@@ -156,7 +155,7 @@ unittest_sources = \
 	tests/unittests/test_cmd_disconnect.c tests/unittests/test_cmd_disconnect.h \
 	tests/unittests/test_callbacks.c tests/unittests/test_callbacks.h \
 	tests/unittests/test_plugins_disco.c tests/unittests/test_plugins_disco.h \
-	tests/unittests/test_http_download.c tests/unittests/test_http_download.h \
+	tests/unittests/test_http_common.c tests/unittests/test_http_common.h \
 	tests/unittests/unittests.c
 
 functionaltest_sources = \
